pyproject.toml
[project]
name = "ocr-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langgraph>=1.0.6",
    "neo4j>=6.1.0",
    "pinecone>=8.0.0",
    "pinecone-text>=0.11.0",
    "rapidfuzz>=3.14.3",
    "sentence-transformers>=5.2.0",
]
